services/yolo/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ultralytics==8.1.0
pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
